qrcode scanner renamed to dynamic feature
