+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/features/DynamicFeatures.kt
@@ -14,6 +14,6 @@ fun DynamicFeature.Companion.qrcodeScanner() = DynamicFeature(
     moduleName = "qrcodescanner",
     titleRes = R.string.title_qrcodescanner,
     descRes = R.string.desc_qrcodescanner,
-    featureLibApiClass = "com.github.klee0kai.thekey.feature.qrcodescanner.FeatureApiImpl",
+    featureLibApiClass = "com.github.klee0kai.thekey.dynamic.qrcodescanner.QRCodeScannerImpl",
 )
