+++ dynamic_qrcodescanner/.gitignore
