+++ dynamic_qrcodescanner/build.gradle.kts
