+++ dynamic_qrcodescanner/src/main/AndroidManifest.xml
