+++ dynamic_qrcodescanner/src/main/kotlin/com/github/klee0kai/thekey/dynamic/qrcodescanner/QRCodeScannerImpl.kt
@@ -1,0 +1,16 @@
+package com.github.klee0kai.thekey.dynamic.qrcodescanner
+
+import com.github.klee0kai.thekey.app.di.AppComponent
+import com.github.klee0kai.thekey.app.features.model.FeatureLibApi
+import com.github.klee0kai.thekey.dynamic.qrcodescanner.di.AndroidHelpersModuleQRExt
+
+@Suppress("unused")
+class QRCodeScannerImpl : FeatureLibApi {
+
+    override fun AppComponent.initDI() {
+        initAndroidHelpersModule(
+            AndroidHelpersModuleQRExt(androidHelpersFactory())
+        )
+    }
+
+}
