+++ dynamic_qrcodescanner/src/main/res/values/strings.xml
