--- qrcodescanner/src/main/kotlin/com/github/klee0kai/thekey/feature/qrcodescanner/FeatureApi.kt
@@ -1,16 +1,0 @@
-package com.github.klee0kai.thekey.feature.qrcodescanner
-
-import com.github.klee0kai.thekey.app.di.AppComponent
-import com.github.klee0kai.thekey.app.features.model.FeatureLibApi
-import com.github.klee0kai.thekey.feature.qrcodescanner.di.AndroidHelpersModuleQRExt
-
-@Suppress("unused")
-class FeatureApiImpl : FeatureLibApi {
-
-    override fun AppComponent.initDI() {
-        initAndroidHelpersModule(
-            AndroidHelpersModuleQRExt(androidHelpersFactory())
-        )
-    }
-
-}
