+++ tkcore/.run/app_mobile.run.xml
@@ -9,7 +9,7 @@
     <option name="ALL_USERS" value="false" />
     <option name="ALWAYS_INSTALL_WITH_PM" value="true" />
     <option name="CLEAR_APP_STORAGE" value="false" />
-    <option name="DYNAMIC_FEATURES_DISABLED_LIST" value="TheKey.dynamicfeature2" />
+    <option name="DYNAMIC_FEATURES_DISABLED_LIST" value="TheKey.dynamic_qrcodescanner" />
     <option name="ACTIVITY_EXTRA_FLAGS" value="" />
     <option name="MODE" value="default_activity" />
     <option name="CLEAR_LOGCAT" value="false" />
